[ENH] Added editing function [CLN] Move gempy functions to modules and regions
