+++ gempy/api_modules/io.py
@@ -6,8 +6,180 @@ when some of the data files are missing
 import pandas as pn
 import numpy as np
 
+# region Save
+from gempy import Model, create_model, init_data
+from gempy.utils.meta import _setdoc
 
 
+@_setdoc(Model.save_model_pickle.__doc__)
+def save_model_to_pickle(model: Model, path=None):
+    model.save_model_pickle(path)
+    return True
+
+
+@_setdoc(Model.save_model.__doc__)
+def save_model(model: Model, name=None, path=None):
+    try:
+        model.grid.topography.topo = None
+    except AttributeError:
+        pass
+    model.save_model(name, path)
+    return True
+
+
+@_setdoc(Model.load_model_pickle.__doc__)
+def load_model_pickle(path):
+    """
+    Read InputData object from python pickle.
+
+    Args:
+       path (str): path where save the pickle
+
+    Returns:
+        :class:`Model`
+
+    """
+    return Model.load_model_pickle(path)
+
+
+def load_model(name, path=None, recompile=False):
+    """
+    Loading model saved with model.save_model function.
+
+    Args:
+        name: name of folder with saved files
+        path (str): path to folder directory
+        recompile (bool): if true, theano functions will be recompiled
+
+    Returns:
+        :class:`Model`
+
+    """
+    # TODO: Divide each dataframe in its own function
+    # TODO: Include try except in case some of the datafiles is missing
+    #
+
+    if not path:
+        path = './'
+    path = f'{path}/{name}'
+
+    # create model with extent and resolution from csv - check
+    geo_model = create_model()
+    init_data(geo_model, np.load(f'{path}/{name}_extent.npy'), np.load(f'{path}/{name}_resolution.npy'))
+    # rel_matrix = np.load()
+    # set additonal data
+    geo_model.additional_data.kriging_data.df = pn.read_csv(f'{path}/{name}_kriging_data.csv', index_col=0,
+                                                            dtype={'range': 'float64', '$C_o$': 'float64',
+                                                                   'drift equations': object,
+                                                                   'nugget grad': 'float64',
+                                                                   'nugget scalar': 'float64'})
+
+    geo_model.additional_data.kriging_data.str2int_u_grade()
+
+    geo_model.additional_data.options.df = pn.read_csv(f'{path}/{name}_options.csv', index_col=0,
+                                                       dtype={'dtype': 'category', 'output': 'category',
+                                                              'theano_optimizer': 'category', 'device': 'category',
+                                                              'verbosity': object})
+    geo_model.additional_data.options.df['dtype'].cat.set_categories(['float32', 'float64'], inplace=True)
+    geo_model.additional_data.options.df['theano_optimizer'].cat.set_categories(['fast_run', 'fast_compile'],
+                                                                                inplace=True)
+    geo_model.additional_data.options.df['device'].cat.set_categories(['cpu', 'cuda'], inplace=True)
+    geo_model.additional_data.options.df['output'].cat.set_categories(['geology', 'gradients'], inplace=True)
+
+    # do series properly - this needs proper check
+    geo_model.series.df = pn.read_csv(f'{path}/{name}_series.csv', index_col=0,
+                                      dtype={'order_series': 'int32', 'BottomRelation': 'category'})
+    series_index = pn.CategoricalIndex(geo_model.series.df.index.values)
+    # geo_model.series.df.index = pn.CategoricalIndex(series_index)
+    geo_model.series.df.index = series_index
+    geo_model.series.df['BottomRelation'].cat.set_categories(['Erosion', 'Onlap', 'Fault'], inplace=True)
+    try:
+        geo_model.series.df['isActive']
+    except KeyError:
+        geo_model.series.df['isActive'] = False
+
+    cat_series = geo_model.series.df.index.values
+
+    # do faults properly - check
+    geo_model.faults.df = pn.read_csv(f'{path}/{name}_faults.csv', index_col=0,
+                                      dtype={'isFault': 'bool', 'isFinite': 'bool'})
+    geo_model.faults.df.index = series_index
+
+    # # do faults relations properly - this is where I struggle
+    geo_model.faults.faults_relations_df = pn.read_csv(f'{path}/{name}_faults_relations.csv', index_col=0)
+    geo_model.faults.faults_relations_df.index = series_index
+    geo_model.faults.faults_relations_df.columns = series_index
+
+    geo_model.faults.faults_relations_df.fillna(False, inplace=True)
+
+    # do surfaces properly
+    surf_df = pn.read_csv(f'{path}/{name}_surfaces.csv', index_col=0,
+                          dtype={'surface': 'str', 'series': 'category',
+                                 'order_surfaces': 'int64', 'isBasement': 'bool', 'id': 'int64',
+                                 'color': 'str'})
+    c_ = surf_df.columns[~(surf_df.columns.isin(geo_model.surfaces._columns_vis_drop))]
+    geo_model.surfaces.df[c_] = surf_df[c_]
+    geo_model.surfaces.df['series'].cat.reorder_categories(np.asarray(geo_model.series.df.index),
+                                                           ordered=False, inplace=True)
+    geo_model.surfaces.sort_surfaces()
+
+    geo_model.surfaces.colors.generate_colordict()
+    geo_model.surfaces.df['series'].cat.set_categories(cat_series, inplace=True)
+
+    try:
+        geo_model.surfaces.df['isActive']
+    except KeyError:
+        geo_model.surfaces.df['isActive'] = False
+
+    cat_surfaces = geo_model.surfaces.df['surface'].values
+
+    # do orientations properly, reset all dtypes
+    geo_model.orientations.df = pn.read_csv(f'{path}/{name}_orientations.csv', index_col=0,
+                                            dtype={'X': 'float64', 'Y': 'float64', 'Z': 'float64',
+                                                   'X_r': 'float64', 'Y_r': 'float64', 'Z_r': 'float64',
+                                                   'dip': 'float64', 'azimuth': 'float64', 'polarity': 'float64',
+                                                   'surface': 'category', 'series': 'category',
+                                                   'id': 'int64', 'order_series': 'int64'})
+    geo_model.orientations.df['surface'].cat.set_categories(cat_surfaces, inplace=True)
+    geo_model.orientations.df['series'].cat.set_categories(cat_series, inplace=True)
+
+    # do surface_points properly, reset all dtypes
+    geo_model.surface_points.df = pn.read_csv(f'{path}/{name}_surface_points.csv', index_col=0,
+                                              dtype={'X': 'float64', 'Y': 'float64', 'Z': 'float64',
+                                                     'X_r': 'float64', 'Y_r': 'float64', 'Z_r': 'float64',
+                                                     'surface': 'category', 'series': 'category',
+                                                     'id': 'int64', 'order_series': 'int64'})
+    geo_model.surface_points.df['surface'].cat.set_categories(cat_surfaces, inplace=True)
+    geo_model.surface_points.df['series'].cat.set_categories(cat_series, inplace=True)
+
+    # Code to add smooth columns for models saved before gempy 2.0bdev4
+    try:
+        geo_model.surface_points.df['smooth']
+    except KeyError:
+        geo_model.surface_points.df['smooth'] = 1e-7
+
+    try:
+        geo_model.orientations.df['smooth']
+    except KeyError:
+        geo_model.orientations.df['smooth'] = 0.01
+
+    # update structure from loaded input
+    geo_model.additional_data.structure_data.update_structure_from_input()
+    geo_model.rescaling.rescale_data()
+    geo_model.update_from_series()
+    geo_model.update_from_surfaces()
+    geo_model.update_structure()
+
+    if recompile is True:
+        from gempy.api_modules.setters import set_interpolator
+        set_interpolator(geo_model, verbose=[0])
+
+    return geo_model
+# endregion
+
+
+# From here down it needs to be reworked to make load more reasonable
+# ----------------
 
 def load_kriging_data(geo_model, path, name):
     geo_model.additional_data.kriging_data.df = pn.read_csv(
@@ -15,20 +187,26 @@ def load_kriging_data(geo_model, path, name):
         dtype={'range': 'float64', '$C_o$': 'float64', 'drift equations': object,
                'nugget grad': 'float64', 'nugget scalar': 'float64'})
 
+
 def load_options(geo_model, path, name):
-    geo_model.additional_data.options.df = pn.read_csv(f'{path}/{name}_options.csv', index_col=0,
-                                            dtype={'dtype': 'category', 'output': 'category',
-                                            'theano_optimizer': 'category', 'device': 'category',
-                                            'verbosity': object})
-    geo_model.additional_data.options.df['dtype'].cat.set_categories(['float32', 'float64'], inplace=True)
-    geo_model.additional_data.options.df['theano_optimizer'].cat.set_categories(['fast_run', 'fast_compile'], inplace=True)
+    geo_model.additional_data.options.df = pn.read_csv(
+        f'{path}/{name}_options.csv', index_col=0,
+        dtype={'dtype': 'category', 'output': 'category',
+               'theano_optimizer': 'category', 'device': 'category',
+               'verbosity': object})
+    geo_model.additional_data.options.df['dtype'].cat.set_categories(
+        ['float32', 'float64'], inplace=True)
+    geo_model.additional_data.options.df['theano_optimizer'].cat.set_categories(
+        ['fast_run', 'fast_compile'],
+        inplace=True)
     geo_model.additional_data.options.df['device'].cat.set_categories(['cpu', 'cuda'], inplace=True)
     geo_model.additional_data.options.df['output'].cat.set_categories(['geology', 'gradients'], inplace=True)
+
 
 def load_series(geo_model, path, name):
     # do series properly - this needs proper check
     geo_model.series.df = pn.read_csv(f'{path}/{name}_series.csv', index_col=0,
-                                            dtype={'order_series': 'int32', 'BottomRelation': 'category'})
+                                      dtype={'order_series': 'int32', 'BottomRelation': 'category'})
     series_index = pn.CategoricalIndex(geo_model.series.df.index.values)
     # geo_model.series.df.index = pn.CategoricalIndex(series_index)
     geo_model.series.df.index = series_index
@@ -40,12 +218,11 @@ def load_series(geo_model, path, name):
 def load_faults(geo_model, path, name):
     # do faults properly - check
     geo_model.faults.df = pn.read_csv(f'{path}/{name}_faults.csv', index_col=0,
-                                            dtype={'isFault': 'bool', 'isFinite': 'bool'})
+                                      dtype={'isFault': 'bool', 'isFinite': 'bool'})
     geo_model.faults.df.index = series_index
 
 
 def load_faults_relations(geo_model, path, name):
-
     # do faults relations properly - this is where I struggle
     geo_model.faults.faults_relations_df = pn.read_csv(f'{path}/{name}_faults_relations.csv', index_col=0)
     geo_model.faults.faults_relations_df.index = series_index
@@ -55,18 +232,16 @@ def load_faults_relations(geo_model, path, name):
 
 
 def load_surfaces(geo_model, path, name):
-
     # do surfaces properly
     geo_model.surfaces.df = pn.read_csv(f'{path}/{name}_surfaces.csv', index_col=0,
-                                            dtype={'surface': 'str', 'series': 'category',
-                                                   'order_surfaces': 'int64', 'isBasement': 'bool', 'id': 'int64'})
+                                        dtype={'surface': 'str', 'series': 'category',
+                                               'order_surfaces': 'int64', 'isBasement': 'bool', 'id': 'int64'})
     geo_model.surfaces.df['series'].cat.set_categories(cat_series, inplace=True)
 
     cat_surfaces = geo_model.surfaces.df['surface'].values
 
 
 def load_orientations(geo_model, path, name):
-
     # do orientations properly, reset all dtypes
     geo_model.orientations.df = pn.read_csv(f'{path}/{name}_orientations.csv', index_col=0,
                                             dtype={'X': 'float64', 'Y': 'float64', 'Z': 'float64',
@@ -79,7 +254,6 @@ def load_orientations(geo_model, path, name):
 
 
 def load_surface_points(geo_model, path, name):
-
     # do surface_points properly, reset all dtypes
     geo_model.surface_points.df = pn.read_csv(f'{path}/{name}_surface_points.csv', index_col=0,
                                               dtype={'X': 'float64', 'Y': 'float64', 'Z': 'float64',
@@ -94,7 +268,6 @@ def load_surface_points(geo_model, path, name):
 
 
 def load_solutions(geo_model, path, name):
-
     # load solutions in npy files
     geo_model.solutions.lith_block = np.load(f'{path}/{name}_lith_block.npy')
     geo_model.solutions.scalar_field_lith = np.load(f"{path}/{name}_scalar_field_lith.npy")
