+++ gempy/core/model.py
@@ -11,7 +11,7 @@ from gempy.core.data import AdditionalData, Faults, Grid, MetaData, Orientations
     Surfaces, Options, Structure, KrigingParameters
 from gempy.core.solution import Solution
 from gempy.core.interpolator import InterpolatorModel, InterpolatorGravity
-from gempy.utils.meta import setdoc, setdoc_pro
+from gempy.utils.meta import _setdoc, _setdoc_pro
 import gempy.utils.docstring as ds
 from gempy.plot.decorators import *
 
@@ -19,9 +19,9 @@ pn.options.mode.chained_assignment = None
 
 
 # TODO rename to ImplicitCoKriging
-@setdoc_pro([Grid.__doc__, Faults.__doc__, Series.__doc__, Surfaces.__doc__, SurfacePoints.__doc__,
-             Orientations.__doc__, RescaledData.__doc__, AdditionalData.__doc__, InterpolatorModel.__doc__,
-             Solution.__doc__])
+@_setdoc_pro([Grid.__doc__, Faults.__doc__, Series.__doc__, Surfaces.__doc__, SurfacePoints.__doc__,
+              Orientations.__doc__, RescaledData.__doc__, AdditionalData.__doc__, InterpolatorModel.__doc__,
+              Solution.__doc__])
 class DataMutation(object):
     """
     This class handles all the mutation of an object belonging to model and the update of every single object depend
@@ -83,9 +83,9 @@ class DataMutation(object):
 
         return idx
 
-    @setdoc_pro([AdditionalData.update_structure.__doc__, InterpolatorModel.set_theano_shared_structure.__doc__,
-                 InterpolatorModel.modify_results_matrices_pro.__doc__,
-                 InterpolatorModel.modify_results_weights.__doc__])
+    @_setdoc_pro([AdditionalData.update_structure.__doc__, InterpolatorModel.set_theano_shared_structure.__doc__,
+                  InterpolatorModel.modify_results_matrices_pro.__doc__,
+                  InterpolatorModel.modify_results_weights.__doc__])
     def update_structure(self, update_theano=None, update_series_is_active=True, update_surface_is_active=True):
         """Update python and theano structure parameters.
 
@@ -172,8 +172,8 @@ class DataMutation(object):
         # TODO this should go to the api and let call all different grid types
         raise NotImplementedError
 
-    @setdoc(Grid.create_regular_grid.__doc__)
-    @setdoc_pro([ds.extent, ds.resolution])
+    @_setdoc(Grid.create_regular_grid.__doc__)
+    @_setdoc_pro([ds.extent, ds.resolution])
     def set_regular_grid(self, extent, resolution):
         """
         Set a regular grid, rescale data and initialize theano solutions.
@@ -200,8 +200,8 @@ class DataMutation(object):
         print(f'Active grids: {self.grid.grid_types[self.grid.active_grids]}')
         return self.grid
 
-    @setdoc(Grid.create_custom_grid.__doc__, )
-    @setdoc_pro(ds.coord)
+    @_setdoc(Grid.create_custom_grid.__doc__, )
+    @_setdoc_pro(ds.coord)
     def set_custom_grid(self, custom_grid):
         """
         Set custom grid, rescale gird and initialize theano solutions. foo
@@ -224,7 +224,7 @@ class DataMutation(object):
         return self.grid
 
     @plot_set_topography
-    @setdoc(Grid.create_topography.__doc__)
+    @_setdoc(Grid.create_topography.__doc__)
     def set_topography(self, source='random', **kwargs):
         """
         Create a topography grid and activate it.
@@ -235,7 +235,7 @@ class DataMutation(object):
         print(f'Active grids: {self.grid.grid_types[self.grid.active_grids]}')
         return self.grid
 
-    @setdoc(Grid.create_centered_grid.__doc__)
+    @_setdoc(Grid.create_centered_grid.__doc__)
     def set_centered_grid(self, centers, radius, resolution=None):
         if self.grid.centered_grid is None:
             self.grid.create_centered_grid(centers, radius, resolution=resolution)
@@ -247,7 +247,7 @@ class DataMutation(object):
         print(f'Active grids: {self.grid.grid_types[self.grid.active_grids]}')
         return self.grid
 
-    @setdoc(Grid.create_section_grid.__doc__)
+    @_setdoc(Grid.create_section_grid.__doc__)
     def set_section_grid(self, section_dict):
         # TODO being able to change the regular grid associated to the section grid
         if self.grid.sections is None:
@@ -271,14 +271,14 @@ class DataMutation(object):
         """
         raise NotImplementedError
 
-    @setdoc([Series.set_bottom_relation.__doc__], indent=False)
+    @_setdoc([Series.set_bottom_relation.__doc__], indent=False)
     def set_bottom_relation(self, series: Union[str, list], bottom_relation: Union[str, list]):
         """"""
         self.series.set_bottom_relation(series, bottom_relation)
         self.interpolator.set_theano_shared_relations()
         return self.series
 
-    @setdoc(Series.add_series.__doc__, indent=False)
+    @_setdoc(Series.add_series.__doc__, indent=False)
     def add_series(self, series_list: Union[str, list], reset_order_series=True):
         """ Add series, update the categories dependet on them and reset the flow control.
         """
@@ -289,7 +289,7 @@ class DataMutation(object):
         self.interpolator.set_flow_control()
         return self.series
 
-    @setdoc(Series.delete_series.__doc__, indent=False)
+    @_setdoc(Series.delete_series.__doc__, indent=False)
     def delete_series(self, indices: Union[str, list], refactor_order_series=True,
                       remove_surfaces=False, remove_data=False):
         """ Delete series, update the categories dependet on them and reset the flow control.
@@ -312,7 +312,7 @@ class DataMutation(object):
         self.interpolator.set_flow_control()
         return self.series
 
-    @setdoc(Series.rename_series.__doc__, indent=False)
+    @_setdoc(Series.rename_series.__doc__, indent=False)
     def rename_series(self, new_categories: Union[dict, list]):
         """Rename series and update the categories dependet on them."""
         self.series.rename_series(new_categories)
@@ -321,7 +321,7 @@ class DataMutation(object):
         self.orientations.df['series'].cat.rename_categories(new_categories, inplace=True)
         return self.series
 
-    @setdoc(Series.modify_order_series.__doc__, indent=False)
+    @_setdoc(Series.modify_order_series.__doc__, indent=False)
     def modify_order_series(self, new_value: int, idx: str):
         """Modify order of the series. Reorder categories of the link Surfaces, sort surface (reset the basement layer)
         remap the Series and Surfaces to the corrspondent dataframes, sort Geometric objects, update structure and
@@ -345,7 +345,7 @@ class DataMutation(object):
         self.update_structure()
         return self.series
 
-    @setdoc(Series.reset_order_series.__doc__, indent=False)
+    @_setdoc(Series.reset_order_series.__doc__, indent=False)
     def reorder_series(self, new_categories: Iterable[str]):
         """Reorder series. Reorder categories of the link Surfaces, sort surface (reset the basement layer)
         remap the Series and Surfaces to the corrspondent dataframes, sort Geometric objects, update structure and
@@ -373,7 +373,7 @@ class DataMutation(object):
     def set_fault_object(self):
         pass
 
-    @setdoc([Faults.set_is_fault.__doc__], indent=False)
+    @_setdoc([Faults.set_is_fault.__doc__], indent=False)
     def set_is_fault(self, series_fault: Union[str, list] = None, toggle: bool = False, offset_faults=False,
                      change_color: bool = True, twofins = False):
         """
@@ -412,7 +412,7 @@ class DataMutation(object):
         self.update_structure(update_theano='matrices')
         return self.faults
 
-    @setdoc([Faults.set_is_finite_fault.__doc__], indent=False)
+    @_setdoc([Faults.set_is_finite_fault.__doc__], indent=False)
     def set_is_finite_fault(self, series_fault=None, toggle: bool = True):
         """ """
         s = self.faults.set_is_finite_fault(series_fault, toggle)  # change df in Fault obj
@@ -420,7 +420,7 @@ class DataMutation(object):
         self.interpolator.set_theano_shared_is_finite()
         return s
 
-    @setdoc([Faults.set_fault_relation.__doc__], indent=False)
+    @_setdoc([Faults.set_fault_relation.__doc__], indent=False)
     def set_fault_relation(self, rel_matrix):
         """"""
         self.faults.set_fault_relation(rel_matrix)
@@ -441,7 +441,7 @@ class DataMutation(object):
         """
         raise NotImplementedError
 
-    @setdoc(Surfaces.add_surface.__doc__, indent=False)
+    @_setdoc(Surfaces.add_surface.__doc__, indent=False)
     def add_surfaces(self, surface_list: Union[str, list], update_df=True):
         self.surfaces.add_surface(surface_list, update_df)
         self.surface_points.df['surface'].cat.add_categories(surface_list, inplace=True)
@@ -449,7 +449,7 @@ class DataMutation(object):
         self.update_structure()
         return self.surfaces
 
-    @setdoc(Surfaces.delete_surface.__doc__, indent=False)
+    @_setdoc(Surfaces.delete_surface.__doc__, indent=False)
     def delete_surfaces(self, indices: Union[str, Iterable[str]], update_id=True, remove_data=True):
         """
         @TODO When implemeted activate geometric data, change remove data to False by default
@@ -489,7 +489,7 @@ class DataMutation(object):
 
         return self.surfaces
 
-    @setdoc(Surfaces.rename_surfaces.__doc__, indent=False)
+    @_setdoc(Surfaces.rename_surfaces.__doc__, indent=False)
     def rename_surfaces(self, to_replace: Union[dict], **kwargs):
 
         self.surfaces.rename_surfaces(to_replace, **kwargs)
@@ -497,7 +497,7 @@ class DataMutation(object):
         self.orientations.df['surface'].cat.rename_categories(to_replace, inplace=True)
         return self.surfaces
 
-    @setdoc(Surfaces.modify_order_surfaces.__doc__, indent=False)
+    @_setdoc(Surfaces.modify_order_surfaces.__doc__, indent=False)
     def modify_order_surfaces(self, new_value: int, idx: int, series_name: str = None):
         """"""
         self.surfaces.modify_order_surfaces(new_value, idx, series_name)
@@ -510,19 +510,19 @@ class DataMutation(object):
         self.update_structure()
         return self.surfaces
 
-    @setdoc(Surfaces.add_surfaces_values.__doc__, indent=False)
+    @_setdoc(Surfaces.add_surfaces_values.__doc__, indent=False)
     def add_surface_values(self,  values_array: Iterable,
                            properties_names: Iterable[str] = np.empty(0)):
         self.surfaces.add_surfaces_values(values_array, properties_names)
         self.update_structure(update_theano='matrices')
         return self.surfaces
 
-    @setdoc(Surfaces.delete_surface_values.__doc__, indent=False)
+    @_setdoc(Surfaces.delete_surface_values.__doc__, indent=False)
     def delete_surface_values(self, properties_names: list):
         self.delete_surface_values(properties_names)
         return self.surfaces
 
-    @setdoc(Surfaces.modify_surface_values.__doc__, indent=False)
+    @_setdoc(Surfaces.modify_surface_values.__doc__, indent=False)
     def modify_surface_values(self, idx, properties_names, values):
         self.surfaces.modify_surface_values(idx, properties_names, values)
         return self.surfaces
@@ -531,7 +531,7 @@ class DataMutation(object):
         self.surfaces.set_surfaces_values(values_array, properties_names)
         return self.surfaces
 
-    @setdoc([Surfaces.map_series.__doc__], indent=False)
+    @_setdoc([Surfaces.map_series.__doc__], indent=False)
     def map_series_to_surfaces(self, mapping_object: Union[dict, pn.Categorical] = None,
                                set_series=True, sort_geometric_data: bool = True, remove_unused_series=True,
                                twofins=False):
@@ -604,7 +604,7 @@ class DataMutation(object):
         possible_candidates = np.array(possible_candidates)
         return possible_candidates[np.isin(possible_candidates, table.columns)][0]
 
-    @setdoc(SurfacePoints.set_surface_points.__doc__, indent=False, position='beg')
+    @_setdoc(SurfacePoints.set_surface_points.__doc__, indent=False, position='beg')
     def set_surface_points(self, table: pn.DataFrame, **kwargs):
         """
         Args:
@@ -644,7 +644,7 @@ class DataMutation(object):
     #    self.additional_data.update_structure()
       #  self.additional_data.update_default_kriging()
 
-    @setdoc(Orientations.set_orientations.__doc__, indent=False, position='beg')
+    @_setdoc(Orientations.set_orientations.__doc__, indent=False, position='beg')
     def set_orientations(self, table: pn.DataFrame, **kwargs):
         """
         Args:
@@ -686,8 +686,8 @@ class DataMutation(object):
         self.update_structure()
        # self.additional_data.update_default_kriging()
 
-    @setdoc_pro(ds.recompute_rf)
-    @setdoc(SurfacePoints.add_surface_points.__doc__, indent=False, position='beg')
+    @_setdoc_pro(ds.recompute_rf)
+    @_setdoc(SurfacePoints.add_surface_points.__doc__, indent=False, position='beg')
     @plot_add_surface_points
     def add_surface_points(self, X, Y, Z, surface, idx: Union[int, Iterable[int]] = None,
                            recompute_rescale_factor=False):
@@ -712,7 +712,7 @@ class DataMutation(object):
 
         return self.surface_points, idx
 
-    @setdoc(SurfacePoints.del_surface_points.__doc__, indent=False, position='beg')
+    @_setdoc(SurfacePoints.del_surface_points.__doc__, indent=False, position='beg')
     @plot_delete_surface_points
     def delete_surface_points(self, idx: Union[int, Iterable[int]]):
         self.surface_points.del_surface_points(idx)
@@ -726,8 +726,8 @@ class DataMutation(object):
         self.delete_surface_points(self.surface_points.df.index[select])
         return True
 
-    @setdoc_pro(ds.recompute_rf)
-    @setdoc(SurfacePoints.modify_surface_points.__doc__, indent=False, position='beg')
+    @_setdoc_pro(ds.recompute_rf)
+    @_setdoc(SurfacePoints.modify_surface_points.__doc__, indent=False, position='beg')
     @plot_move_surface_points
     def modify_surface_points(self, indices: Union[int, list], recompute_rescale_factor=False, **kwargs):
         """
@@ -767,8 +767,8 @@ class DataMutation(object):
     def set_orientations_object(self, orientations: Orientations, update_model=True):
         raise NotImplementedError
 
-    @setdoc_pro(ds.recompute_rf)
-    @setdoc(Orientations.add_orientation.__doc__, indent=False, position='beg')
+    @_setdoc_pro(ds.recompute_rf)
+    @_setdoc(Orientations.add_orientation.__doc__, indent=False, position='beg')
     @plot_add_orientation
     def add_orientations(self,  X, Y, Z, surface, pole_vector: Iterable = None,
                          orientation: Iterable = None, idx=None, recompute_rescale_factor=False):
@@ -794,14 +794,14 @@ class DataMutation(object):
 
         return self.orientations, idx
 
-    @setdoc(Orientations.del_orientation.__doc__, indent=False, position='beg')
+    @_setdoc(Orientations.del_orientation.__doc__, indent=False, position='beg')
     @plot_delete_orientations
     def delete_orientations(self, idx: Union[list, int]):
         self.orientations.del_orientation(idx)
         self.update_structure(update_theano='weights')
         return self.orientations
 
-    @setdoc(Orientations.modify_orientations.__doc__, indent=False, position='beg')
+    @_setdoc(Orientations.modify_orientations.__doc__, indent=False, position='beg')
     @plot_move_orientations
     def modify_orientations(self, idx: list, **kwargs):
 
@@ -819,7 +819,7 @@ class DataMutation(object):
     # endregion
 
     # region Options
-    @setdoc(Options.modify_options.__doc__, indent=False, position='beg')
+    @_setdoc(Options.modify_options.__doc__, indent=False, position='beg')
     def modify_options(self, attribute, value):
         self.additional_data.options.modify_options(attribute, value)
         warnings.warn('You need to recompile the Theano code to make it the changes in options.')
@@ -827,7 +827,7 @@ class DataMutation(object):
     # endregion
 
     # region Kriging
-    @setdoc(KrigingParameters.modify_kriging_parameters.__doc__, indent=False, position='beg')
+    @_setdoc(KrigingParameters.modify_kriging_parameters.__doc__, indent=False, position='beg')
     def modify_kriging_parameters(self, attribute, value, **kwargs):
         self.additional_data.kriging_data.modify_kriging_parameters(attribute, value, **kwargs)
         self.interpolator.set_theano_shared_kriging()
@@ -838,7 +838,7 @@ class DataMutation(object):
     # endregion
 
     # region rescaling
-    @setdoc(RescaledData.modify_rescaling_parameters.__doc__, indent=False, position='beg')
+    @_setdoc(RescaledData.modify_rescaling_parameters.__doc__, indent=False, position='beg')
     def modify_rescaling_parameters(self, attribute, value):
         self.additional_data.rescaling_data.modify_rescaling_parameters(attribute, value)
         self.additional_data.rescaling_data.rescale_data()
@@ -892,7 +892,7 @@ class DataMutation(object):
         self.update_from_surfaces()
         return self.surfaces
 
-    @setdoc_pro(ds.extent)
+    @_setdoc_pro(ds.extent)
     def set_extent(self, extent: Iterable):
         """
         Set project extent
@@ -1005,7 +1005,7 @@ class DataMutation(object):
         return True
 
     # region Theano interface
-    @setdoc(InterpolatorModel.__doc__)
+    @_setdoc(InterpolatorModel.__doc__)
     def set_theano_graph(self, interpolator: InterpolatorModel,  update_structure=True, update_kriging=True):
         """ Pass a theano graph of a Interpolator instance other than the Model compose
 
@@ -1024,7 +1024,7 @@ class DataMutation(object):
         self.update_to_interpolator()
         return True
 
-    @setdoc(InterpolatorModel.__doc__)
+    @_setdoc(InterpolatorModel.__doc__)
     def set_theano_function(self, interpolator: InterpolatorModel,
                             update_structure=True, update_kriging=True):
         """
@@ -1145,7 +1145,7 @@ class DataMutation(object):
 
 
 # TODO rename to Project. With DEP time
-@setdoc([MetaData.__doc__, DataMutation.__doc__], indent=False)
+@_setdoc([MetaData.__doc__, DataMutation.__doc__], indent=False)
 class Model(DataMutation, ABC):
     """ Container class of all objects that constitute a GemPy model.
 
@@ -1256,7 +1256,7 @@ class Model(DataMutation, ABC):
 
         return True
 
-    @setdoc([SurfacePoints.read_surface_points.__doc__, Orientations.read_orientations.__doc__])
+    @_setdoc([SurfacePoints.read_surface_points.__doc__, Orientations.read_orientations.__doc__])
     def read_data(self, source_i=None, source_o=None, add_basement=True, **kwargs):
         """
         Read data from a csv, or directly supplied dataframes
