+++ test/conftest.py
@@ -29,17 +29,34 @@ def interpolator_magnetics():
 
 
 @pytest.fixture(scope='session')
-def one_fault_model(interpolator):
-    model = gp.create_data([0, 2000, 0, 2000, 0, 2000], [50, 50, 50],
+def one_fault_model_no_interp():
+    """This only makes sense for running small test fast"""
+    model = gp.create_data('one_fault_model', [0, 2000, 0, 2000, 0, 2000], [50, 50, 50],
                            path_o=input_path2 + 'tut_chapter1/simple_fault_model_orientations.csv',
                            path_i=input_path2 + 'tut_chapter1/simple_fault_model_points.csv')
 
     # Assigning series to surface as well as their order (timewise)
     gp.map_series_to_surfaces(model, {"Fault_Series": 'Main_Fault',
                                       "Strat_Series": ('Sandstone_2', 'Siltstone',
-                                      'Shale', 'Sandstone_1')},
-                  )
+                                                       'Shale', 'Sandstone_1')},
+                              )
     model.set_is_fault(['Fault_Series'])
+    return model
+
+
+@pytest.fixture(scope='session')
+def one_fault_model(one_fault_model_no_interp, interpolator):
+    # model = gp.create_data([0, 2000, 0, 2000, 0, 2000], [50, 50, 50],
+    #                        path_o=input_path2 + 'tut_chapter1/simple_fault_model_orientations.csv',
+    #                        path_i=input_path2 + 'tut_chapter1/simple_fault_model_points.csv')
+    #
+    # # Assigning series to surface as well as their order (timewise)
+    # gp.map_series_to_surfaces(model, {"Fault_Series": 'Main_Fault',
+    #                                   "Strat_Series": ('Sandstone_2', 'Siltstone',
+    #                                   'Shale', 'Sandstone_1')},
+    #               )
+    # model.set_is_fault(['Fault_Series'])
+    model = one_fault_model_no_interp
     model.set_theano_function(interpolator)
 
     return model
@@ -53,7 +70,7 @@ def one_fault_model_solution(one_fault_model):
 
 @pytest.fixture(scope='session')
 def unconformity_model(interpolator):
-    geo_model = gp.create_data(
+    geo_model = gp.create_data('unconformity_model',
         [0, 1000, 0, 1000, 0, 1000],
         resolution=[50, 50, 50],
         path_o=input_path2 + "jan_models/model6_orientations.csv",
