+++ test/test_api/test_api.py
@@ -1,0 +1,9 @@
+import gempy as gp
+
+
+def test_edit_1(one_fault_model_no_interp):
+    model = one_fault_model_no_interp
+    rpn = gp.edit(model, 'surfaces', 'add_surface', surface_list=['foo'])
+
+    assert rpn.df.iloc[-1, 0] == 'foo'
+    assert model.surfaces.df.iloc[-1, 0]
