+++ test/test_core/test_grids/test_grid.py
@@ -21,7 +21,7 @@ class TestGrid:
                        resolution=[50, 50, 50])
 
     def test_section_grid(self):
-        geo_data = gp.create_data([0, 1000, 0, 1000, 0, 1000], resolution=[10, 10, 10])
+        geo_data = gp.create_data('section_grid', [0, 1000, 0, 1000, 0, 1000], resolution=[10, 10, 10])
         geo_data.set_topography()
         section_dict = {'section1': ([0, 0], [1000, 1000], [100, 80]),
                         'section2': ([800, 0], [800, 1000], [150, 100]),
